The @jsonrpc_method no longer implicitly registers methods with jsonrpc_site.
This is a little backward incompatible. But easily fixable.

Instead, the decorated methods must now be imported in the Django
project's urls module, and registered explicitly with the jsonrpc_site
there.

The old behaviour, where you would import the views module and the methods
would get registered implicitly, reeked of side effects.

The new behaviour is better because:
1. explicit is better than implicit
2. you can now control, in your Django project's urls module, which
   JSONRPCSite gets which methods. In fact, it is now easy to register
   one method with multiple JSONRPCSites, if you need to.

Another backward-incompatible change is the signature of
JSONRPCSite.register(). The JSON-RPC method name is now an optional
argument, coming *after* the method itself. That's because when you use
the @jsonrpc_method decorator, the JSON-RPC method name is already
attached to the decorated method as the json_method attribute anyway.
However, I doubt this particular change affects anybody as nobody really
needed to call JSONRPCSite.register() explicitly before.